# apps/vc-api/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/vc-api/credentials/issue": {
      "post": {
        "operationId": "VcApiController_issueCredential",
        "summary": "",
        "description": "Issues a credential and returns it in the response body. Conforms to https://w3c-ccg.github.io/vc-api/issuer.html",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/IssueCredentialDto" } }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/VerifiableCredentialDto" } }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/BadRequestErrorResponseDto" } }
            }
          }
        },
        "tags": ["vc-api"]
      }
    },
    "/v1/vc-api/credentials/verify": {
      "post": {
        "operationId": "VcApiController_verifyCredential",
        "summary": "",
        "description": "Verify a credential. Conforms to https://w3c-ccg.github.io/vc-api/#verify-credential",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/VerifyCredentialDto" } }
          }
        },
        "responses": {
          "200": {
            "description": "Verifiable Credential successfully verified",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/VerificationResultDto" } }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/BadRequestErrorResponseDto" } }
            }
          }
        },
        "tags": ["vc-api"]
      }
    },
    "/v1/vc-api/presentations/prove/authentication": {
      "post": {
        "operationId": "VcApiController_proveAuthenticationPresentation",
        "summary": "",
        "description": "Issue a DIDAuth presentation that authenticates a DID.\nNot a part of VC-API? Maybe there is a DID Auth spec though?\nA NON-STANDARD endpoint currently.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AuthenticateDto" } } }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/VerifiablePresentationDto" } }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/BadRequestErrorResponseDto" } }
            }
          }
        },
        "tags": ["vc-api"]
      }
    },
    "/v1/vc-api/presentations/from": {
      "post": {
        "operationId": "VcApiController_presentationFrom",
        "summary": "",
        "description": "Creates a Presentation without Proof by passing in the Presentation Definition, selected Verifiable Credentials (TODO: and an optional holder (DID)).\nThe presentation contains the [presentation submission](https://identity.foundation/presentation-exchange/#presentation-submission) data that the verifier can use.",
        "parameters": [],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/PresentationDto" } }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/BadRequestErrorResponseDto" } }
            }
          }
        },
        "tags": ["vc-api"]
      }
    },
    "/v1/vc-api/presentations/prove": {
      "post": {
        "operationId": "VcApiController_provePresentation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/ProvePresentationDto" } }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/VerifiablePresentationDto" } }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/BadRequestErrorResponseDto" } }
            }
          }
        },
        "tags": ["vc-api"]
      }
    },
    "/v1/vc-api/presentations/verify": {
      "post": {
        "operationId": "VcApiController_verifyPresentation",
        "summary": "",
        "description": "Verify a presentation. Conforms to https://w3c-ccg.github.io/vc-api/#verify-presentation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/VerifyPresentationDto" } }
          }
        },
        "responses": {
          "200": {
            "description": "Verifiable Presentation successfully verified!",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/VerificationResultDto" } }
            }
          },
          "400": { "description": "Invalid or malformed input" }
        },
        "tags": ["vc-api"]
      }
    },
    "/v1/vc-api/exchanges": {
      "post": {
        "operationId": "VcApiController_createExchange",
        "summary": "",
        "description": "Allows the creation of a new exchange by providing the credential query and interaction endpoints\nA NON-STANDARD endpoint currently.\n\nSimilar to https://gataca-io.github.io/vui-core/#/Presentations/post_api_v2_presentations",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/ExchangeDefinitionDto" } }
          }
        },
        "responses": {
          "201": { "description": "" },
          "400": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/BadRequestErrorResponseDto" } }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ConflictErrorResponseDto" } }
            }
          }
        },
        "tags": ["vc-api"]
      }
    },
    "/v1/vc-api/exchanges/{exchangeId}": {
      "post": {
        "operationId": "VcApiController_initiateExchange",
        "summary": "",
        "description": "Initiates an exchange of information.\nhttps://w3c-ccg.github.io/vc-api/#initiate-exchange",
        "parameters": [
          { "name": "exchangeId", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ExchangeResponseDto" } }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/BadRequestErrorResponseDto" } }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/NotFoundErrorResponseDto" } }
            }
          }
        },
        "tags": ["vc-api"]
      }
    },
    "/v1/vc-api/exchanges/{exchangeId}/{transactionId}": {
      "put": {
        "operationId": "VcApiController_continueExchange",
        "summary": "",
        "description": "Receives information related to an existing exchange.\nhttps://w3c-ccg.github.io/vc-api/#continue-exchange",
        "parameters": [
          { "name": "exchangeId", "required": true, "in": "path", "schema": { "type": "string" } },
          { "name": "transactionId", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/VerifiablePresentationDto" } }
          }
        },
        "responses": {
          "200": {
            "description": "Verifiable Presentation successfully submitted and verified",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ExchangeResponseDto" } }
            }
          },
          "202": {
            "description": "Verifiable Presentation successfully submitted. Further review in progress."
          },
          "400": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/BadRequestErrorResponseDto" } }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/NotFoundErrorResponseDto" } }
            }
          }
        },
        "tags": ["vc-api"]
      },
      "get": {
        "operationId": "VcApiController_getTransaction",
        "summary": "",
        "description": "Get exchange transaction by id\nA NON-STANDARD endpoint currently.\nSimilar to https://identitycache.energyweb.org/api/#/Claims/ClaimController_getByIssuerDid",
        "parameters": [
          { "name": "exchangeId", "required": true, "in": "path", "schema": { "type": "string" } },
          { "name": "transactionId", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/GetTransactionDto" } }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/BadRequestErrorResponseDto" } }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/NotFoundErrorResponseDto" } }
            }
          }
        },
        "tags": ["vc-api"]
      }
    },
    "/v1/vc-api/exchanges/{exchangeId}/{transactionId}/review": {
      "post": {
        "operationId": "VcApiController_addSubmissionReview",
        "summary": "",
        "description": "Update a transaction review\nA NON-STANDARD endpoint currently.\nSimilar to https://github.com/energywebfoundation/ssi-hub/blob/8b860e7cdae4e1b1aa75afeab8b9df7ab26befbb/src/modules/claim/claim.controller.ts#L80",
        "parameters": [
          { "name": "exchangeId", "required": true, "in": "path", "schema": { "type": "string" } },
          { "name": "transactionId", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/SubmissionReviewDto" } }
          }
        },
        "responses": {
          "201": { "description": "" },
          "400": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/BadRequestErrorResponseDto" } }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/NotFoundErrorResponseDto" } }
            }
          }
        },
        "tags": ["vc-api"]
      }
    }
  },
  "info": { "title": "VC-API", "description": "Sample VC-API", "version": "0.1", "contact": {} },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "BadRequestErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": { "type": "number", "example": 400 },
          "message": {
            "oneOf": [
              { "type": "string", "example": "error" },
              { "type": "array", "items": { "type": "string" }, "example": ["error 1", "error 2"] }
            ]
          },
          "error": { "type": "string", "example": "Bad Request" }
        },
        "required": ["statusCode", "message", "error"]
      },
...
      "IssueCredentialDto": {
        "type": "object",
        "properties": {
          "credential": { "$ref": "#/components/schemas/CredentialDto" },
          "options": { "$ref": "#/components/schemas/IssueOptionsDto" }
        },
        "required": ["credential", "options"]
      },
      "VerifiableCredentialDto": {
        "type": "object",
        "properties": {
          "@context": {
            "type": "array",
            "description": "The JSON-LD context of the credential.",
            "items": { "oneOf": [{ "type": "string" }, { "type": "object" }] }
          },
          "id": { "type": "string", "description": "The ID of the credential." },
          "type": {
            "description": "The JSON-LD type of the credential.",
            "type": "array",
            "items": { "type": "string" }
          },
          "issuer": { "type": "string", "description": "A JSON-LD Verifiable Credential Issuer." },
          "issuanceDate": {
            "type": "string",
            "description": "The issuanceDate",
            "example": "2022-09-21T11:49:03.205Z"
          },
          "expirationDate": {
            "type": "string",
            "description": "The expirationDate",
            "example": "2023-09-21T11:49:03.205Z"
          },
          "credentialSubject": { "type": "object", "description": "The subject" },
          "proof": { "type": "object", "description": "A JSON-LD Linked Data proof." }
        },
        "required": ["@context", "id", "type", "issuer", "issuanceDate", "credentialSubject", "proof"]
      },
...
      "VerifyCredentialDto": {
        "type": "object",
        "properties": {
          "vc": {
            "description": "A JSON-LD Verifiable Credential with a proof. https://w3c-ccg.github.io/vc-api/issuer.html#operation/issueCredential",
            "allOf": [{ "$ref": "#/components/schemas/VerifiableCredentialDto" }]
          },
          "options": {
            "description": "Parameters for verifying a verifiable credential or a verifiable presentation https://w3c-ccg.github.io/vc-api/verifier.html#operation/verifyCredential https://w3c-ccg.github.io/vc-api/verifier.html#operation/verifyPresentation",
            "allOf": [{ "$ref": "#/components/schemas/VerifyOptionsDto" }]
          }
        },
        "required": ["vc", "options"]
      },
      "VerificationResultDto": {
        "type": "object",
        "properties": {
          "checks": { "description": "The checks performed", "type": "array", "items": { "type": "string" } },
          "warnings": { "description": "Warnings", "type": "array", "items": { "type": "string" } },
          "errors": { "description": "Errors", "type": "array", "items": { "type": "string" } }
        },
        "required": ["checks", "warnings", "errors"]
      },
...
      "AuthenticateDto": {
        "type": "object",
        "properties": {
          "did": { "type": "string" },
          "options": { "$ref": "#/components/schemas/ProvePresentationOptionsDto" }
        },
        "required": ["did", "options"]
      },
      "VerifiablePresentationDto": {
        "type": "object",
        "properties": {
          "@context": {
            "type": "array",
            "description": "The JSON-LD context of the presentation.",
            "items": { "oneOf": [{ "type": "string" }, { "type": "object" }] }
          },
          "id": {
            "type": "string",
            "description": "The ID of the presentation. The id property is optional and MAY be used to provide a unique identifier for the presentation. https://www.w3.org/TR/vc-data-model/#presentations-0"
          },
          "type": {
            "description": "The JSON-LD type of the presentation.",
            "type": "array",
            "items": { "type": "string" }
          },
          "holder": {
            "type": "string",
            "description": "The holder - will be ignored if no proof is present since there is no proof of authority over the credentials"
          },
          "verifiableCredential": {
            "description": "The Verifiable Credentials",
            "type": "array",
            "items": { "$ref": "#/components/schemas/VerifiableCredentialDto" }
          },
          "proof": { "type": "object", "description": "A JSON-LD Linked Data proof." }
        },
        "required": ["@context", "type", "proof"]
      },
      "PresentationDto": {
        "type": "object",
        "properties": {
          "@context": {
            "type": "array",
            "description": "The JSON-LD context of the presentation.",
            "items": { "oneOf": [{ "type": "string" }, { "type": "object" }] }
          },
          "id": {
            "type": "string",
            "description": "The ID of the presentation. The id property is optional and MAY be used to provide a unique identifier for the presentation. https://www.w3.org/TR/vc-data-model/#presentations-0"
          },
          "type": {
            "description": "The JSON-LD type of the presentation.",
            "type": "array",
            "items": { "type": "string" }
          },
          "holder": {
            "type": "string",
            "description": "The holder - will be ignored if no proof is present since there is no proof of authority over the credentials"
          },
          "verifiableCredential": {
            "description": "The Verifiable Credentials",
            "type": "array",
            "items": { "$ref": "#/components/schemas/VerifiableCredentialDto" }
          }
        },
        "required": ["@context", "type"]
      },
      "ProvePresentationDto": {
        "type": "object",
        "properties": {
          "presentation": { "$ref": "#/components/schemas/PresentationDto" },
          "options": { "$ref": "#/components/schemas/ProvePresentationOptionsDto" }
        },
        "required": ["presentation", "options"]
      },
      "VerifyPresentationDto": {
        "type": "object",
        "properties": {
          "vp": {
            "description": "A JSON-LD Verifiable Credential with a proof. https://w3c-ccg.github.io/vc-api/issuer.html#operation/issueCredential",
            "allOf": [{ "$ref": "#/components/schemas/VerifiablePresentationDto" }]
          },
          "options": {
            "description": "Parameters for verifying a verifiable credential or a verifiable presentation https://w3c-ccg.github.io/vc-api/verifier.html#operation/verifyCredential https://w3c-ccg.github.io/vc-api/verifier.html#operation/verifyPresentation",
            "allOf": [{ "$ref": "#/components/schemas/VerifyOptionsDto" }]
          }
        },
        "required": ["vp", "options"]
      },
...
      "ExchangeDefinitionDto": {
        "type": "object",
        "properties": {
          "exchangeId": {
            "type": "string",
            "description": "The id of the exchange. It must be unique in the context of each VC API instance."
          },
          "interactServices": {
            "description": "The Interact Service Definitions are related to the Interaction Types of the Verifiable Presentation Request (VPR) specification.\nHowever, as it is a configuration object, it not identical to a VPR interact services.\nIt can be see as the input data that the application uses to generate VPR interact services during the exchanges.",
            "type": "array",
            "items": { "$ref": "#/components/schemas/ExchangeInteractServiceDefinitionDto" }
          },
          "query": {
            "description": "Defines requests for data in the Verifiable Presentation",
            "type": "array",
            "items": { "$ref": "#/components/schemas/VpRequestQueryDto" }
          },
          "isOneTime": {
            "type": "boolean",
            "description": "Indicates whether or not the exchange should only be used once.\n\nIf wanting to ensure that the exchange will only be used once, this should be true.\nThis could be useful to, for example, associate the exchange is a specific instance of a business process.\n\nIf wanting to generated an exchange endpoint that can be reused, this should be false.\nThis could be useful to, for example, generate an exchange url that be put on a QR code sticker and distributed"
          },
          "callback": {
            "description": "An array of \"callbacks\" that will be used by VC API to send notifications on the status/result of the exchange.",
            "type": "array",
            "items": { "$ref": "#/components/schemas/CallbackConfigurationDto" }
          }
        },
        "required": ["exchangeId", "interactServices", "query", "isOneTime", "callback"]
      },
      "ConflictErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": { "type": "number", "example": 404 },
          "message": { "type": "string" },
          "error": { "type": "string", "example": "Not Found" }
        },
        "required": ["statusCode", "message", "error"]
      },
...
      "ExchangeResponseDto": {
        "type": "object",
        "properties": {
          "errors": {
            "description": "Any errors encountered during exchange",
            "type": "array",
            "items": { "type": "string" }
          },
          "vpRequest": {
            "description": "Verifiable Presentation Request.\nShould conform to VP-Request specification.\nWill be returned if a VP is required to obtain more information from requester\nMay not be returned if no further information is required (for example, at the end of the workflow)",
            "allOf": [{ "$ref": "#/components/schemas/VpRequestDto" }]
          },
          "vp": {
            "description": "If it is an issuance response, then a vp may be provided",
            "allOf": [{ "$ref": "#/components/schemas/VerifiablePresentationDto" }]
          },
          "processingInProgress": {
            "type": "boolean",
            "description": "True if an exchange submission is currently being processed or reviewed asyncronously"
          }
        },
        "required": ["errors", "processingInProgress"]
      },
...
    }
  }
}